src/main/java/ch/hesge/algo/K.java
package ch.hesge.algo;

import ch.hesge.algo.model.Company;
import ch.hesge.algo.model.Department;
import ch.hesge.algo.model.Employee;

import java.util.Set;

public class K {

    /**
     * En utilisant uniquement l'API streams,
     * associer tous les Employee qui n'ont pas de Company à la Company dans le département ENGINEERING.
     *
     * @param company   Company qui recrute
     * @param employees Collection d'Employee
     */
    public void hireUnemployedEmployees(Company company, Set<Employee> employees) {
        employees.stream()
                .filter(employee -> !employee.getCompany().isPresent())
                .forEach(employee -> company.hire(employee, Department.ENGINEERING));
    }
}
